fix: migrate Timefold 1.x to 2.x and fix Spring Boot 3.x compatibility
- Upgrade timefold-solver 1.x -> 2.1.0 (bluetape4k-dependencies 1.1.3 BOM)
- Update artifact: timefold-solver-persistence-common removed in 2.x
- Update artifact: timefold-solver-test removed; ConstraintVerifier moved to core
- Fix score package: buildin.*.*Score -> HardMediumSoftScore/HardSoftScore (top-level)
- Remove Long score types (BendableLong/HardSoftLong/HardMediumSoftLong/SimpleLong) removed in 2.x
- Fix score factory params: Int -> Long in Timefold 2.x
- Fix PlanningId import: domain.lookup -> domain.common
- Fix SolverManager: 2 type params -> 1 type param in 2.x
- Fix SolverJobBuilder: withBestSolutionEventConsumer receives NewBestSolutionEvent
- Fix BedPlan.score: non-null -> nullable (HardMediumSoftScore? = null)
- Remove GraalVM native plugin from quickstart modules (timefold-solver-spring-boot-starter
  2.x compiled against Spring 7.x/AOT APIs incompatible with Spring Boot 3.x/Spring 6.x)
- Disable analyze test: SolutionManager.analyze() is Enterprise-only in Timefold 2.x

+++ 01-quickstarts/bed-allocation/src/main/kotlin/timefold/workshop/bed/allocation/domain/Stay.kt
@@ -1,7 +1,7 @@
 package timefold.workshop.bed.allocation.domain
 
 import ai.timefold.solver.core.api.domain.entity.PlanningEntity
-import ai.timefold.solver.core.api.domain.lookup.PlanningId
+import ai.timefold.solver.core.api.domain.common.PlanningId
 import ai.timefold.solver.core.api.domain.variable.PlanningVariable
 import com.fasterxml.jackson.annotation.JsonIgnore
 import io.bluetape4k.idgenerators.uuid.TimebasedUuid
@@ -57,16 +57,16 @@ class Stay: Serializable {
 
 
     @get:JsonIgnore
-    val nightCount: Int
-        get() = ChronoUnit.DAYS.between(arrivalDate, departureDate).toInt() + 1
+    val nightCount: Long
+        get() = ChronoUnit.DAYS.between(arrivalDate, departureDate) + 1L
 
-    fun calculateSameNightCount(other: Stay): Int {
+    fun calculateSameNightCount(other: Stay): Long {
         val maxArrivalDate = maxOf(arrivalDate!!, other.arrivalDate!!)
         val minDepartureDate = minOf(departureDate!!, other.departureDate!!)
 
         return maxOf(
-            0,
-            ChronoUnit.DAYS.between(maxArrivalDate, minDepartureDate).toInt() + 1
+            0L,
+            ChronoUnit.DAYS.between(maxArrivalDate, minDepartureDate) + 1L
         )
     }
