+++ exposed/jdbc-examples/src/test/kotlin/timefold/workshop/persistence/exposed/SimpleScoreTest.kt
@@ -1,6 +1,6 @@
 package timefold.workshop.persistence.exposed
 
-import ai.timefold.solver.core.api.score.buildin.simple.SimpleScore
+import ai.timefold.solver.core.api.score.SimpleScore
 import io.bluetape4k.exposed.dao.entityToStringBuilder
 import io.bluetape4k.exposed.dao.idEquals
 import io.bluetape4k.exposed.dao.idHashCode
@@ -50,7 +50,7 @@ class SimpleScoreTest: AbstractScoreExposedTest() {
         withTables(testDB, T1) {
 
             val name = faker.name().name()
-            val simpleScore = SimpleScore.of(faker.random().nextInt(0, 100))
+            val simpleScore = SimpleScore.of(faker.random().nextLong(0, 100))
 
             val id = T1.insertAndGetId {
                 it[T1.name] = name
@@ -75,7 +75,7 @@ class SimpleScoreTest: AbstractScoreExposedTest() {
         withTables(testDB, T1) {
             val saved = E1.new {
                 this.name = faker.name().name()
-                this.simpleScore = SimpleScore.of(faker.random().nextInt(0, 100))
+                this.simpleScore = SimpleScore.of(faker.random().nextLong(0, 100))
             }
             log.debug { "saved=$saved" }
             flushCache()
